Reference

​

Contents

​

Reference

Contents

Index

​

Index

​

PartiallySeparableSolvers.ModTrustRegionPartitionedData.Counter

PartiallySeparableSolvers.Mod_PQN.PartitionedDataTRPQN

NLPModels.hprod

NLPModels.hprod!

PartiallySeparableSolvers.ModPartitionedMethods.PTRUNK

PartiallySeparableSolvers.ModTrustRegionPartitionedData.compute_ratio

PartiallySeparableSolvers.ModTrustRegionPartitionedData.partitionedTrunk

PartiallySeparableSolvers.Mod_PQN.build_PartitionedDataTRPQN

PartiallySeparableSolvers.Mod_PQN.partitionedMulOp!

PartiallySeparableSolvers.Mod_PQN.update_PQN

PartiallySeparableSolvers.Mod_PQN.update_PQN!

PartiallySeparableSolvers.Mod_ab_partitioned_data.evaluate_grad_part_data

PartiallySeparableSolvers.Mod_ab_partitioned_data.evaluate_grad_part_data!

PartiallySeparableSolvers.Mod_ab_partitioned_data.evaluate_obj_part_data

PartiallySeparableSolvers.Mod_ab_partitioned_data.evaluate_obj_part_data!

PartiallySeparableSolvers.Mod_ab_partitioned_data.evaluate_y_part_data!

PartiallySeparableSolvers.Mod_ab_partitioned_data.product_part_data_x

PartiallySeparableSolvers.Mod_ab_partitioned_data.product_part_data_x!

PartiallySeparableSolvers.Mod_ab_partitioned_data.update_nlp!

​

PartiallySeparableSolvers.ModPartitionedMethods.PTRUNK — Method

stats = PTRUNK(nlp::AbstractNLPModel; name = :plse, kwargs...)

PTRUNK (partitioned update solver) return a stats::GenericExecutionStats from a partitioned quasi-Newton trust-region method. Several variants are available via the optional argument name. Each variant approximate the element-Hessians differently, some use dense matrices while others use linear-operators. Variants with dense element-Hessian approximations:

PBFGS with name=:pbfgs, every element-Hessian approximation is updated with BFGS;

PSR1 with name=:psr1, every element-Hessian approximation is updated with SR1;

PSE with name=:pse, every element-Hessian approximation is updated with BFGS or SR1 if the curvature condition doesn't hold.

Variants with linear-operator element-Hessian approximations:

PLBFGS with name=:plbfgs, every element-Hessian approximations is a LBFGS operator;

PLSR1 with name=:plsr1, every element-Hessian approximations is a LSR1 operator;

PLSE with name=:plse, by default, every element-Hessian approximations is a LBFGS operator as long as the curvature condition holds, otherwise it becomes a LSR1 operator.

PartiallySeparableSolvers.ModTrustRegionPartitionedData.Counter — Type

Counter

Substitute to NLPModels.Counters since the partitioned methods don't relie (for now) on PartiallySeparableNLPModels. It has fields:

neval_obj::Int: count objective evaluations;

neval_grad::Int: count gradient computations;

neval_Hprod::Int: count Hessian-approximation-vector products.

PartiallySeparableSolvers.ModTrustRegionPartitionedData.compute_ratio — Method

ρₖ = compute_ratio(x::AbstractVector{T}, fₖ::T, sₖ::Vector{T}, part_data::PartitionedData, B::AbstractLinearOperator{T}, gₖ::AbstractVector{T}; cpt::Counter = Counter(0, 0, 0))

Compute the ratio between the actual loss and the expected loss using part_data, the current point x and the step s. g_k must be the gradient at x and B the linear-operator paired to part_data.

PartiallySeparableSolvers.ModTrustRegionPartitionedData.partitionedTrunk — Method

stats = partitionedTrunk(nlp::AbstractNLPModel, part_data::PartitionedData; max_eval::Int = 10000, max_iter::Int = 10000, start_time::Float64 = time(), max_time::Float64 = 30.0, ϵ::Float64 = 1e-6, name = part_data.name, name_method::String = "Trust-region " * String(name), kwargs...)

Produce a GenericExecutionStats for a partitioned quasi-Newton trust-region method. It requires the partitioned structures of part_data::PartitionedDate paired with an nlp model. The counter nlp.counters are updated with the informations of ModTrustRegionPartitionedData.Counter to ease the definition of a GenericExecutionStats.

PartiallySeparableSolvers.Mod_PQN.PartitionedDataTRPQN — Type

PartitionedDataTRPQN{G, T <: Number, P <: Part_mat{T}} <: PartitionedData

Gather the structures required to run a partitioned quasi-Newton trust-region method. PartitionedDataTRPQN has fields:

n: the size of the problem;

N: the number of element functions;

vec_elt_fun: a ElementFunction vector, of size N;

M: the number of distinct element-function expression trees;

vec_elt_complete_expr_tree: a Complete_expr_tree vector, of size M;

element_expr_tree_table: a vector of size M, the i-th element element_expr_tree_table[i]::Vector{Int} informs which element functions use the vec_elt_complete_expr_tree[i] expression tree;

index_element_tree: a vector of size N where each component indicates which Complete_expr_tree from vec_elt_complete_expr_tree is used for the corresponding element;

vec_compiled_element_gradients: the vector gathering the compiled tapes for every element gradient evaluations;

x: the current point;

v: a temporary vector;

s: the current step;

pg: the partitioned gradient;

pv: a temporary partitioned vector;

py: the partitioned gradient difference;

ps: the partitioned step;

phv: the partitioned Hessian-vector product;

pB: the partitioned matrix (main memory cost);

fx: the current value of the objective function;

name: the name of partitioned quasi-Newton update performed at each iteration:

name=:pbfgs: every element-Hessian approximation is updated with BFGS;

name=:psr1: every element-Hessian approximation is updated with SR1;

name=:pse: every element-Hessian approximation is updated with BFGS if the curvature condition holds, or with SR1 otherwise;

name=:pcs: each element-Hessian approximation with BFGS if it is classified as convex, or with SR1 otherwise;

name=:plbfgs: every element-Hessian approximations is a LBFGS operator;

name=:plsr1: every element-Hessian approximations is a LSR1 operator;

name=:plse: by default, every element-Hessian approximations is a LBFGS operator as long as the curvature condition holds, otherwise it becomes a LSR1 operator.

PartiallySeparableSolvers.Mod_PQN.build_PartitionedDataTRPQN — Method

partitioneddata_tr_pqn = build_PartitionedDataTRPQN(expr_tree, n)

Return the structure required to run a partitioned quasi-Newton trust-region method. It finds the partially-separable structure of an expression tree expr_tree representing f(x) = ∑fᵢ(xᵢ). Then it allocates the partitioned structures required. To define properly the sparse matrix of the partitioned matrix we need the size of the problem: n.

PartiallySeparableSolvers.Mod_PQN.partitionedMulOp! — Method

partitionedMulOp!(pd_pqn::PartitionedDataTRPQN{G, T, P}, res, v, α, β) where {G, T, P}

Partitioned 5-arg mul! for pd_pqn using the partitioned matrix and partitioned vectors to destribute and collect the result of element matrix-vector products.

PartiallySeparableSolvers.Mod_PQN.update_PQN! — Method

update_PQN!(pd_pqn::PartitionedDataTRPQN{G, T, P}, s::Vector{T})
update_PQN!(pd_pqn::PartitionedDataTRPQN{G, T, P}, x::Vector{T}, s::Vector{T})

Perform the partitioned quasi-Newton update given the current point x and the step s. When x is omitted, update_PQN! consider that pd_pqn has the current point in pdpqn.x. Moreover, it assumes that the partitioned gradient atxis already computed inpdpqn.pg`.

PartiallySeparableSolvers.Mod_PQN.update_PQN — Method

B = update_PQN(pd_pqn::PartitionedDataTRPQN{G, T, P}, x::Vector{T}, s::Vector{T};

Perform the partitioned quasi-Newton update given the current point x and the step s.

PartiallySeparableSolvers.Mod_ab_partitioned_data.update_nlp! — Method

update_nlp!(pd_pqn::PartitionedDataTRPQN{G, T, P}, s::Vector{T})
update_nlp!(pd_pqn::PartitionedDataTRPQN{G, T, P}, x::Vector{T}, s::Vector{T})

Perform the partitioned quasi-Newton update given the current point x and the step s. When x is omitted, update_PQN! consider that pd_pqn has the current point in pdpqn.x. Moreover, it assumes that the partitioned gradient atxis already computed inpdpqn.pg`.

NLPModels.hprod! — Method

hprod!(part_data::PartitionedData, x::AbstractVector, v::AbstractVector, hv::AbstractVector)

Set hv::Vector to the product between the partitioned hessian ∇²f(x) and the vector v.

NLPModels.hprod — Method

hv = hprod(part_data::PartitionedData, x::AbstractVector, v::AbstractVector)

Return hv::Vector the product between the partitioned hessian ∇²f(x) and the vector v.

PartiallySeparableSolvers.Mod_ab_partitioned_data.evaluate_grad_part_data! — Method

evaluate_grad_part_data!(g::AbstractVector{Y}, part_data::PartitionedData, x::AbstractVector{Y}) where {Y <: Number}
evaluate_grad_part_data!(part_data::PartitionedData)

Evaluate the gradient at x in place. It computes and store the element gradients in part_data.pg and accumulate their contributions in g. When g and x are omitted, consider that part_data.g and part_data.x are g and x, respectively.

PartiallySeparableSolvers.Mod_ab_partitioned_data.evaluate_grad_part_data — Method

gradient = evaluate_grad_part_data(part_data::PartitionedData, x::AbstractVector{Y}) where {Y <: Number}

Return the gradient vector g at x. After the computation of the element gradients (stored in part_data.pg), g is built by accumulating their contributions.

PartiallySeparableSolvers.Mod_ab_partitioned_data.evaluate_obj_part_data! — Method

evaluate_obj_part_data!(part_data::PartitionedData)

Compute the partially separable objective function, as a sum of element functions, at part_data.x. Its objective value is stored in part_data.fx .

PartiallySeparableSolvers.Mod_ab_partitioned_data.evaluate_obj_part_data — Method

fx = evaluate_obj_part_data(part_data::PartitionedData, x::AbstractVector{Y}) where {Y <: Number}

Return the partially separable objective at x.

PartiallySeparableSolvers.Mod_ab_partitioned_data.evaluate_y_part_data! — Method

evaluate_y_part_data!(part_data::PartitionedData, x::AbstractVector{Y}, s::AbstractVector{Y}) where {Y <: Number}
evaluate_y_part_data!(part_data::PartitionedData, s::AbstractVector{Y}) where {Y <: Number}

Compute element gradients differences ∇̂fᵢ(x+s)-∇̂fᵢ(x) for each element function. Store the results in part_data.py. When x is ommited, consider part_data.x as x and part_data.pg as the partitioned gradient at x.

PartiallySeparableSolvers.Mod_ab_partitioned_data.product_part_data_x! — Method

product_part_data_x!(res::AbstractVector{Y}, part_data::PartitionedData, v::AbstractVector{Y}) where {Y <: Number}
product_part_data_x!(epv_res::PartitionedStructures.Elemental_pv{Y}, part_data::PartitionedData, epv::PartitionedStructures.Elemental_pv{Y}) where {Y <: Number}
product_part_data_x!(epv_res::PartitionedStructures.Elemental_pv{Y}, pB::T, epv::PartitionedStructures.Elemental_pv{Y}) where {Y <: Number, T <: PartitionedStructures.Part_mat{Y}}

Store the product between the partitioned matrix part_data.pB and the vector v in res. The computation of every element matrix-vector product require two partitioned vectors. The first partitioned vector epv is created from v and the results are stored in the second partitioned vector epv_res which builds the value store by g.

PartiallySeparableSolvers.Mod_ab_partitioned_data.product_part_data_x — Method

Bx = product_part_data_x(part_data::PartitionedData, v::AbstractVector{Y}) where {Y <: Number}

Return the product between the partitioned matrix part_data.pB and v.