Reference

​

Contents

​

Reference

Contents

Index

​

Index

​

PartiallySeparableSolvers.Mod_TR_CG_part_data.Counter

PartiallySeparableSolvers.Mod_TR_CG_part_data.TR_CG_PD

PartiallySeparableSolvers.Mod_TR_CG_part_data.compute_ratio

PartiallySeparableSolvers.Mod_TR_CG_part_data.generic_algorithm_wrapper

PartiallySeparableSolvers.Mod_partitioned_methods.PUS

PartiallySeparableSolvers.Mod_partitioned_methods.get_expr_tree

PartiallySeparableSolvers.solver_TR_CG_Ab_NLP_LO

PartiallySeparableSolvers.solver_TR_CG_Ab_NLP_LO_ges

PartiallySeparableSolvers.upgrade_TR_LO!

​

PartiallySeparableSolvers.solver_TR_CG_Ab_NLP_LO — Method

(x, iter) = solver_TR_CG_Ab_NLP_LO(nlp::AbstractNLPModel, B::AbstractLinearOperator{T}; max_eval::Int = 10000, max_iter::Int = 10000, start_time::Float64 = time(), max_time::Float64 = 30.0,)

A limited-memory quasi-Newton trust-region solver where the subproblems, whose quadratics terms are updated through the linear-operator B, are solved with a conjugate-gradient method (see Krylov.jl). The method return the point x and the number of iterations performed before it reaches the stopping criterias.

PartiallySeparableSolvers.solver_TR_CG_Ab_NLP_LO_ges — Method

ges = solver_TR_CG_Ab_NLP_LO_ges(nlp::AbstractNLPModel, B::AbstractLinearOperator{T}; max_eval::Int = 10000, max_iter::Int = 10000, start_time::Float64 = time(), max_time::Float64 = 30.0) where {T <: Number}

Return a GenericExecutionStats from the quasi-Newton trust-region method solver_TR_CG_Ab_NLP_LO given the linear-operator B.

PartiallySeparableSolvers.upgrade_TR_LO! — Method

Δ = upgrade_TR_LO(pk::Float64, x_k::AbstractVector{T}, s_k::AbstractVector{T}, g_k::AbstractVector{T},  y_k::AbstractVector{T},  B_k::AbstractLinearOperator{T}, nlp::AbstractNLPModel, Δ::Float64; η::Float64 = 1e-3, η1::Float64 = 0.75 ) where {T <: Number}

Update the linear-operator Bₖ and return the trust-region radius Δ depending ρₖ the ratio computed beforehand from the step sₖ.

PartiallySeparableSolvers.Mod_partitioned_methods.PUS — Method

ges = PUS(nlp::AbstractNLPModel; name = :plse, kwargs...)

PUS (partitioned update solver) return a ges::GenericExecutionStats from a partitioned quasi-Newton trust-region method. It can perform several variants, which can be select with the optional argument name. You will perform a:

PBFGS method with name=:pbfgs;

PSR1 method with name=:psr1;

PSE method with name=:pse;

PLBFGS method with name=:plbfgs;

PLSR1 method with name=:plsr1;

PLSE method with name=:plse, by default.

PartiallySeparableSolvers.Mod_partitioned_methods.get_expr_tree — Method

expr_tree, n, x0 = get_expr_tree(nlp::MathOptNLPModel; x0::Vector{T} = copy(nlp.meta.x0), kwargs...) where {T <: Number}
expr_tree, n, x0 = get_expr_tree(adnlp::ADNLPModel; x0::Vector{T} = copy(adnlp.meta.x0), kwargs...) where {T <: Number}

Return the expr_tree, the size n and the initial point x0 from either a MathOptNLPModel or a ADNLPModel.

PartiallySeparableSolvers.Mod_TR_CG_part_data.Counter — Type

Counter

Substitute to NLPModels.Counters since the partitioned methods don't relie (for now) on PartiallySeparableNLPModels. It has fields:

neval_obj::Int: count objective evaluations;

neval_grad::Int: count gradient computations;

neval_Hprod::Int: count Hessian-approximation-vector products.

PartiallySeparableSolvers.Mod_TR_CG_part_data.TR_CG_PD — Method

(x, iter) = TR_CG_PD(part_data::PartiallySeparableNLPModels.PartitionedData; x::AbstractVector = copy(get_x(part_data)), n::Int = get_n(part_data), max_eval::Int = 10000, max_iter::Int = 10000, max_time::Float64 = 30.0, atol::Real = √eps(eltype(x)), rtol::Real = √eps(eltype(x)), start_time::Float64 = time(), η::Float64 = 1e-3, η₁::Float64 = 0.75, # > η Δ::Float64 = 1.0, ϵ::Float64 = 1e-6, ϕ::Float64 = 2.0, ∇f₀::AbstractVector = PartiallySeparableNLPModels.evaluate_grad_part_data(part_data, x), cpt::Counter = Counter(0, 0, 0), iter_print::Int64 = Int(floor(max_iter / 100)), T = eltype(x), verbose = true, kwargs...)

Partitioned quasi-Newton trust-region method apply on the partitioned structures of part_data. The method return the point x and the number of iterations performed before it reaches the stopping criterias.

PartiallySeparableSolvers.Mod_TR_CG_part_data.compute_ratio — Method

ρₖ = compute_ratio(x::AbstractVector{T}, fₖ::T, sₖ::Vector{T}, part_data::PartiallySeparableNLPModels.PartitionedData, B::AbstractLinearOperator{T}, gₖ::AbstractVector{T}; cpt::Counter = Counter(0, 0, 0))

Compute the ratio between the actual loss and the expected loss using part_data, the current point x and the step s. g_k must be the gradient at x and B the linear-operator paired to part_data.

PartiallySeparableSolvers.Mod_TR_CG_part_data.generic_algorithm_wrapper — Method

ges = generic_algorithm_wrapper(nlp::AbstractNLPModel, part_data::PartiallySeparableNLPModels.PartitionedData; max_eval::Int = 10000, max_iter::Int = 10000, start_time::Float64 = time(), max_time::Float64 = 30.0, ϵ::Float64 = 1e-6, name = part_data.name, name_method::String = "Trust-region " * String(name), kwargs...)

Produce a GenericExecutionStats for a partitioned quasi-Newton trust-region method. It requires the partitioned structures of part_data::PartitionedDate paired with an nlp model. The counter nlp.counters are updated with the informations of Mod_TR_CG_part_data.Counter to ease the definition of a GenericExecutionStats.