Reference

​

Contents

​

Reference

Contents

Index

​

Index

​

PartiallySeparableSolvers.alloc_struct_algo

PartiallySeparableSolvers.approx_quad

PartiallySeparableSolvers.compute_ratio

PartiallySeparableSolvers.init_struct_algo!

PartiallySeparableSolvers.other_index

PartiallySeparableSolvers.solver_TR_CG_Ab_NLP_LO

PartiallySeparableSolvers.solver_TR_CG_Ab_NLP_Trunk

PartiallySeparableSolvers.solver_TR_PBFGS!

PartiallySeparableSolvers.update_BFGS!

PartiallySeparableSolvers.update_PBFGS!

PartiallySeparableSolvers.update_PBS!

PartiallySeparableSolvers.update_PSR1!

PartiallySeparableSolvers.update_SPS_BFGS!

PartiallySeparableSolvers.update_SPS_SR1!

PartiallySeparableSolvers.update_SPS_mix_SR1_BFGS!

PartiallySeparableSolvers.update_SR1!

PartiallySeparableSolvers.upgrade_TR_LO!

PartiallySeparableSolvers.upgrade_TR_LO_no_update!

​

PartiallySeparableSolvers.alloc_struct_algo — Method

alloc_struct_algo(obj, n, type)

Alloc the structure needed for the whole Trust Region algorithm, which include the gradients vectors and Hessian approximations, the vectors to store the points xₖ and xₖ₋₁, some constants Δ, η... and the Partially separable structure of the obj function.

PartiallySeparableSolvers.approx_quad — Method

approx_quad(struct_algo, pₖ   )

return the quadratic approximation m(pₖ) = fₖ + gₖᵀpₖ + 1/2.pₖᵀBpₖ. The values of fₖ, gₖ and Bₖ are stored inside struct_algo.

PartiallySeparableSolvers.compute_ratio — Method

compute_ratio(struct_algo, sₖ )

Compute the ratio : (fₖ - fₖ₊₁)/(mₖ(0)-mₖ(sₖ)) , all the data about fₖ, mₖ is stored in struct_algo

PartiallySeparableSolvers.init_struct_algo! — Method

initstructalgo(structalgo, x ) Once the structure is allocated, we can use initstruct to init the structure struct_algo at the point x, which is the initial point of the algorithm. When the initialisation is done we can start the algorithme.

PartiallySeparableSolvers.other_index — Method

other_index(i)

i ∈ {1,2}. We use a trick to avoid useless copy of gradients and Hessians in the structure of the algorithm. So we need and index in the structure of the algorithm. The function otherindex return the other index. The index is an enumerate type. If the index is currently to 1 (fst) in the structure sa then otherindex(sa) return 2 (snd), and in the other case return 1 (fst)

PartiallySeparableSolvers.solver_TR_CG_Ab_NLP_LO — Method

solver_L_SR1_Ab_NLP(nlp, B, x0)

solverLSR1AbNLP is a optimisation method using Trust region with conjugate gradient. nlp is an AbstractNLPModel, B is an AbstractLinearOperator and x0 is an AbstractVector reprensenting the initial point of the method. The method return a tuple: with the final point and the number of iteration performming by the method. Even if the name suggest LSR1, B can be a BFGS Operator.

PartiallySeparableSolvers.solver_TR_CG_Ab_NLP_Trunk — Method

solver_TR_CG_Ab_NLP_Trunk(nlp, B)

solverLSR1AbNLP is a optimisation method using Trust region with conjugate gradient. nlp is an AbstractNLPModel, B is an AbstractLinearOperator and x0 is an AbstractVector reprensenting the initial point of the method. The method return a tuple: with the final point and the number of iteration performming by the method. Even if the name suggest LSR1, B can be a BFGS Operator.

PartiallySeparableSolvers.solver_TR_PBFGS! — Method

solver_TR_PSR1!(model)

Trust region method using the gradient conjugate method and the Partially Separable Structure of the model from the parameters. This method use BFGS approximation.

PartiallySeparableSolvers.update_BFGS! — Method

update_BFGS(Δx, y, B, B_1 )

function that builde the next approximation of the Hessian, which will be stored in B_1. Rhe result is based from Δx, y and B; respectively the difference between 2 points, the difference of the gradient of the associate points, and the previous approximation of the Hessian. The approximation is made according to the BFGS update method.

PartiallySeparableSolvers.update_PBFGS! — Method

update_PBFGS!(struct_algo, B)

This function perform a step of a Trust-Region method using a conjuguate-gragient method to solve the sub-problem of the Trust-Region. B is the LinearOperator needed by the cg (conjuguate-gragient method). struct_algo stored all the data relative to the problem and is modified if step is taken .

PartiallySeparableSolvers.update_PBS! — Method

update_PBS!(struct_algo, B)

This function perform a step of a Trust-Region method using a conjuguate-gragient method to solve the sub-problem of the Trust-Region. B is the LinearOperator needed by the cg (conjuguate-gragient method). struct_algo stored all the data relative to the problem and is modified if step is taken .

PartiallySeparableSolvers.update_PSR1! — Method

update_PSR1!(struct_algo, B)

This function perform a step of a Trust-Region method using a conjuguate-gragient method to solve the sub-problem of the Trust-Region. B is the LinearOperator needed by the cg (conjuguate-gragient method). struct_algo stored all the data relative to the problem and is modified if step is taken .

PartiallySeparableSolvers.update_SPS_BFGS! — Method

update_SPS_BFGS(sps, Bₖ, Bₖ₊₁, yₖ, sₖ)

update the Hessian approximation Bₖ using the SR1 method, according to the sps partially separable structre. To make the update, we need the gradvector y and the vector s. B, B1 and y use structure linked with the partially separable structure stored in sps.

PartiallySeparableSolvers.update_SPS_SR1! — Method

update_SPS_SR1(sps, Bₖ, Bₖ₊₁, yₖ, sₖ)

update the Hessian approximation Bₖ using the SR1 method, according to the sps partially separable structre. To make the update, we need the gradvector y and the vector s. B, B1 and y use structure linked with the partially separable structure stored in sps.

PartiallySeparableSolvers.update_SPS_mix_SR1_BFGS! — Method

update_SPS_mix_SR1_BFGS(sps, Bₖ, Bₖ₊₁, yₖ, sₖ)

The update is based on the gradvector y and the vector s. B, B1 and y use structure linked with the partially separable structure stored in sps. We update the Hessian approximation Bₖ according with partially separable structure. We update each Bᵢₖ,if it is convex we use BFGS rather than SR1.

PartiallySeparableSolvers.update_SR1! — Method

update_BFGS(Δx, y, B, B_1 )

function that builde the next approximation of the Hessian, which will be stored in B_1. The result is based from Δx, y and B; respectively the difference between 2 points, the difference of the gradient of the associate points, and the previous approximation of the Hessian. The approximation is made according to the SR1 update method.

PartiallySeparableSolvers.upgrade_TR_LO! — Method

upgrade_TR_LO(ρₖ, xₖ , sₖ, gₖ, Bₖ, nlp, Δ)

Update the LinearOperator Bₖ using the push! if the ratio ρₖ is accepted by computing xₖ₊₁ = xₖ + sₖ and gₖ₊₁ from xₖ₊₁. Update the radius of the Trust region Δ.

PartiallySeparableSolvers.upgrade_TR_LO_no_update! — Method

upgrade_TR_LO(ρₖ, xₖ , sₖ, gₖ, Bₖ, nlp, Δ)

Update the LinearOperator Bₖ using the push! if the ratio ρₖ is accepted by computing xₖ₊₁ = xₖ + sₖ and gₖ₊₁ from xₖ₊₁. Update the radius of the Trust region Δ.